Admin Update Edit

+++ php/admin/deleteRecord.php
@@ -5,18 +5,30 @@
     }
 </style>
 <?php
+session_start();
 
 error_reporting(E_ALL);
 ini_set('display_errors', 1);
 
-include('../../models/teacher/teacher_class.php');
-include('../../DB.php');
-echo "Delete called<br>";
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
-    echo $_POST["deleteBtn"];
-    echo "<br>Calling class";
-    $ha = new Teacher;
-    $ha->delete($_POST["deleteBtn"]);
-    header('Location: ../database.php');
+
+    // If table to be deleted is teachers==================================
+    if ($_SESSION['tableName'] == 'teachers') {
+        include('../../models/teacher/teacher_class.php');
+        $ha = new Teacher;
+        $ha->delete($_POST["deleteBtn"]);
+        header('Location: ../../php/admin/database.php');
+    }
+    
+    // It table is applicant============================================
+    if ( $_SESSION['tableName'] == 'applicant') {
+        include('../../models/applicant/applicant_class.php');
+        $applicantObj = new applicant();
+        if($applicantObj->delete($_POST['deleteBtn'])){
+            header('Location: ../../php/admin/database.php');
+        } 
+    }
+
+
 }
 ?>
